+++ CHANGELOG.md
@@ -303,6 +303,7 @@ Internal changes:
 * Minor refactoring related to the handling of wandering traders not supporting the baby state.
 * Minor: All type registries (shop types, shop object types, UI types, ..) remember the order of their registered types now. This should have no noticeable effect, other than maybe on the order of command argument completion suggestions.
 * Most block and entity shop object types share the same object ids now. The idea is that this might allow for optimization when doing shopkeeper lookups.
+* Minor refactoring related to the handling of chat inputs when naming shopkeepers.
 
 Config changes:  
 * The default value of the `prevent-shop-creation-item-regular-usage` setting was changed to `true`.